Updated Tag Service Tests

+++ CECS 491B/TrialByFire.Tresearch/TrialByFire.Tresearch.Services/Implementations/TagService.cs
@@ -1,6 +1,9 @@
-﻿using TrialByFire.Tresearch.DAL.Contracts;
+﻿using Microsoft.Extensions.Options;
+using TrialByFire.Tresearch.DAL.Contracts;
 using TrialByFire.Tresearch.Services.Contracts;
+using TrialByFire.Tresearch.Models;
 using TrialByFire.Tresearch.Models.Contracts;
+using TrialByFire.Tresearch.Models.Implementations;
 
 namespace TrialByFire.Tresearch.Services.Implementations
 {
@@ -15,6 +18,8 @@ namespace TrialByFire.Tresearch.Services.Implementations
 
         private IMessageBank _messageBank { get; set; }
 
+        private BuildSettingsOptions _options { get; }
+
         /// <summary>
         ///     public RecoveryClass(sqlDAO, logService, messageBank)
         ///         Constructor for Tag service class
@@ -22,21 +27,47 @@ namespace TrialByFire.Tresearch.Services.Implementations
         /// <param name="sqlDAO"> SQL object to perform changes to database</param>
         /// <param name="logService"> Log service</param>
         /// <param name="messageBank"> Message bank containing status code enumerables</param>
-        public TagService(ISqlDAO sqlDAO, ILogService logService, IMessageBank messageBank)
+        public TagService(ISqlDAO sqlDAO, ILogService logService, IMessageBank messageBank, IOptions<BuildSettingsOptions> options)
         {
             _sqlDAO = sqlDAO;
             _logService = logService;
             _messageBank = messageBank;
+            _options = options.Value;
         }
 
+        /// <summary>
+        ///     AddTagToNodeAsync(nodeIDs, tagName)
+        ///         Adds a tag to node. Node must be owned by user and 
+        /// </summary>
+        /// <param name="nodeIDs"></param>
+        /// <param name="tagName"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
         public async Task<string> AddTagToNodesAsync(List<long> nodeIDs, string tagName, CancellationToken cancellationToken = default(CancellationToken))
         {
             try
             {
                 cancellationToken.ThrowIfCancellationRequested();
+
+                string role = "";
+                if (Thread.CurrentPrincipal.IsInRole(_options.User))
+                    role = _options.User;
+                else if (Thread.CurrentPrincipal.IsInRole(_options.Admin))
+                    role = _options.Admin;
+                else
+                    return await _messageBank.GetMessage(IMessageBank.Responses.unknownRole).ConfigureAwait(false);
+
+                IAccount account = new Account(Thread.CurrentPrincipal.Identity.Name, role);
+
+                Tuple<bool, string> isAuthorized = await OwnsNode(nodeIDs, account, cancellationToken);
+
+                if (!isAuthorized.Item1)
+                    return isAuthorized.Item2;
+
                 if (nodeIDs.Count > 0)
                 {
                     string result = await _sqlDAO.AddTagToNodesAsync(nodeIDs, tagName, cancellationToken);
+
                     if(cancellationToken.IsCancellationRequested && result.Equals(_messageBank.GetMessage(IMessageBank.Responses.generic).Result))
                     {
                         //Perform Rollback
@@ -46,6 +77,7 @@ namespace TrialByFire.Tresearch.Services.Implementations
                         else
                             return _messageBank.GetMessage(IMessageBank.Responses.generic).Result;
                     }
+                    
                     return result;
                 }
                 else
@@ -57,7 +89,7 @@ namespace TrialByFire.Tresearch.Services.Implementations
             }
             catch(Exception ex)
             {
-                return "500: Server: " + ex.Message;
+                return _options.UncaughtExceptionMessage + ex.Message;
             }
         }
 
@@ -66,6 +98,22 @@ namespace TrialByFire.Tresearch.Services.Implementations
             try
             {
                 cancellationToken.ThrowIfCancellationRequested();
+
+                string role = "";
+                if (Thread.CurrentPrincipal.IsInRole(_options.User))
+                    role = _options.User;
+                else if (Thread.CurrentPrincipal.IsInRole(_options.Admin))
+                    role = _options.Admin;
+                else
+                    return await _messageBank.GetMessage(IMessageBank.Responses.unknownRole).ConfigureAwait(false);
+
+                IAccount account = new Account(Thread.CurrentPrincipal.Identity.Name, role);
+
+                Tuple<bool, string> isAuthorized = await OwnsNode(nodeIDs, account, cancellationToken);
+
+                if (!isAuthorized.Item1)
+                    return isAuthorized.Item2;
+
                 if (nodeIDs.Count > 0)
                 {
                     string result = await _sqlDAO.RemoveTagFromNodeAsync(nodeIDs, tagName, cancellationToken);
@@ -89,7 +137,7 @@ namespace TrialByFire.Tresearch.Services.Implementations
             }
             catch (Exception ex)
             {
-                return "500: Server: " + ex.Message;
+                return _options.UncaughtExceptionMessage + ex.Message;
             }
         }
 
@@ -99,6 +147,22 @@ namespace TrialByFire.Tresearch.Services.Implementations
             try
             {
                 cancellationToken.ThrowIfCancellationRequested();
+
+                string role = "";
+                if (Thread.CurrentPrincipal.IsInRole(_options.User))
+                    role = _options.User;
+                else if (Thread.CurrentPrincipal.IsInRole(_options.Admin))
+                    role = _options.Admin;
+                else
+                    return Tuple.Create(new List<string>(), await _messageBank.GetMessage(IMessageBank.Responses.unknownRole).ConfigureAwait(false));
+
+                IAccount account = new Account(Thread.CurrentPrincipal.Identity.Name, role);
+
+                Tuple<bool, string> isAuthorized = await OwnsNode(nodeIDs, account, cancellationToken);
+
+                if (!isAuthorized.Item1)
+                    return Tuple.Create(new List<string>(), isAuthorized.Item2);
+
                 if (nodeIDs.Count > 0)
                 {
                     Tuple<List<string>, string> result = await _sqlDAO.GetNodeTagsAsync(nodeIDs, cancellationToken);
@@ -115,7 +179,33 @@ namespace TrialByFire.Tresearch.Services.Implementations
             }
             catch (Exception ex)
             {
-                return Tuple.Create(tags, "500: Server: " + ex.Message);
+                return Tuple.Create(tags, _options.UncaughtExceptionMessage + ex.Message);
+            }
+        }
+
+        public async Task<Tuple<List<string>, string>> GetNodeTagsDescAsync(List<long> nodeIDs, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            List<string> tags = new List<string>();
+            try
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                if (nodeIDs.Count > 0)
+                {
+                    Tuple<List<string>, string> result = await _sqlDAO.GetNodeTagsDescAsync(nodeIDs, cancellationToken);
+                    if (cancellationToken.IsCancellationRequested)
+                        throw new OperationCanceledException();
+                    return result;
+                }
+                else
+                    return Tuple.Create(tags, _messageBank.GetMessage(IMessageBank.Responses.nodeTagNodeDoesNotExist).Result);
+            }
+            catch (OperationCanceledException)
+            {
+                throw;
+            }
+            catch (Exception ex)
+            {
+                return Tuple.Create(tags, _options.UncaughtExceptionMessage + ex.Message);
             }
         }
 
@@ -142,7 +232,7 @@ namespace TrialByFire.Tresearch.Services.Implementations
             }
             catch (Exception ex)
             {
-                return "500: Server: " + ex.Message;
+                return _options.UncaughtExceptionMessage + ex.Message;
             }
         }
 
@@ -163,7 +253,28 @@ namespace TrialByFire.Tresearch.Services.Implementations
             }
             catch (Exception ex)
             {
-                return Tuple.Create(tags, "500: Server: " + ex.Message);
+                return Tuple.Create(tags, _options.UncaughtExceptionMessage + ex.Message);
+            }
+        }
+
+        public async Task<Tuple<List<string>, string>> GetTagsDescAsync(CancellationToken cancellationToken = default(CancellationToken))
+        {
+            List<string> tags = null;
+            try
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                Tuple<List<string>, string> result = await _sqlDAO.GetTagsDescAsync(cancellationToken);
+                if (cancellationToken.IsCancellationRequested)
+                    throw new OperationCanceledException();
+                return result;
+            }
+            catch (OperationCanceledException)
+            {
+                throw;
+            }
+            catch (Exception ex)
+            {
+                return Tuple.Create(tags, _options.UncaughtExceptionMessage + ex.Message);
             }
         }
 
@@ -190,7 +301,27 @@ namespace TrialByFire.Tresearch.Services.Implementations
             }
             catch (Exception ex)
             {
-                return "500: Server: " + ex.Message;
+                return _options.UncaughtExceptionMessage + ex.Message;
+            }
+        }
+
+        public async Task<Tuple<bool, string>> OwnsNode(List<long> nodeID, IAccount account, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                Tuple<bool, string> result = await _sqlDAO.IsAuthorizedToMakeNodeChangesAsync(nodeID, account, cancellationToken);
+                if (cancellationToken.IsCancellationRequested)
+                    throw new OperationCanceledException();
+                return result;
+            } 
+            catch (OperationCanceledException)
+            {
+                throw;
+            }
+            catch (Exception ex)
+            {
+                return Tuple.Create(false, _options.UncaughtExceptionMessage + ex.Message);
             }
         }
     }
